requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.20
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2==2.8.4
py==1.8.0
pycodestyle==2.5.0
pylint==2.3.1
pymongo==3.8.0
pyparsing==2.4.2
pytest==5.1.2
python-dotenv==0.10.1
redis==3.3.8
requests==2.21.0
six==1.12.0
typed-ast==1.3.4
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==0.15.2
wrapt==1.11.1
zipp==0.6.0
